examples/claude-tool-use/requirements.txt
anthropic
httpx
web3
python-dotenv
